Allow defined metadata that is not used.
This change ensures that metadata defined in the schema does not cause a validation error if it is not used within that schema. A change to the logic in this validation is required in order to support a new feature within author to add and remove metadata for a questionnaire.

The reverse logic should still work as expected. i.e. metadata that is being used without first being defined should still cause an error during validation.