Remove check for placeholders on non-mandatory answers
Runner uses the pattern of an optional question with a
ConfirmationQuestion and later questions using the piped
answer quite often, the change being reverted does not allow
this pattern.

Ideally, the routing path should be worked out in validator
to allow it to check whether the piped answer will only be
displayed if the question has been answered.