Migrate to draft 07 and add ajv Debug Server (#175)
An update of validator to support draft 07, along with the introduction of an ajv based server to be used during debug of schemas. We could split the schema/questionnaire validation between two implementations, but this would make our current workflow even more complex. The errors returned by the ajv implementation are ordered by their path length, which should mean the first error is the most useful as it will have matched on a particular block.

Update to use Python 3.8.0 and black formatting.